% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_table.R
\name{apa_table}
\alias{apa_table}
\title{apa_table}
\usage{
apa_table(table)
}
\arguments{
\item{table}{input data table}
}
\value{
gt object
}
\description{
covert a data table to APA format
}
\details{
black border on the top and bottom of the columns, black border on the bottom of the table, and font is Times size 12
}
