% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_colnames.R
\name{append_prefix_to_colnames}
\alias{append_prefix_to_colnames}
\title{append_prefix_to_colnames}
\usage{
append_prefix_to_colnames(data, prefix)
}
\arguments{
\item{data}{data.frame that contains the column names to change with a prefix}

\item{prefix}{string to add to the beginning of all of the column names of the data.frame}
}
\value{
data.frame with new column names with prefix string appended to the column names
}
\description{
add a string to beginning of all the column names
}
\examples{
df <- append_prefix_to_colnames(carData::Salaries, 'p_')
head(df)
}
\seealso{
Other column name helpers:
\code{\link[=append_colnames]{append_colnames()}},
\code{\link[=append_suffix_to_colnames]{append_suffix_to_colnames()}},
\code{\link[=rename_colnames]{rename_colnames()}},
\code{\link[=tidy_colnames]{tidy_colnames()}}
}
\concept{column name helpers}
\concept{data_wrangle}
