% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_out_path.R
\name{check_out_path}
\alias{check_out_path}
\title{check_out_path}
\usage{
check_out_path(out_path, overwrite = FALSE, warn_only = FALSE)
}
\arguments{
\item{out_path}{path to output file}

\item{overwrite}{booleon to overwrite the output path (default: FALSE)}

\item{warn_only}{boolean to warn instead of stop if file exists and overwrite set to FALSE (default: FALSE)}
}
\value{

}
\description{
\preformatted{
checks to see if output file exists.

1. STOP WITH ERROR if output file exists, overwrite set to FALSE and warn_only set to FALSE
2. WARN if output file exists, overwrite set to FALSE and warn_only set to TRUE
3. WARN if output file exists and overwrite set to TRUE
}
}
\examples{

}
\concept{r_helper}
