% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_all_models.R
\name{compare_all_models}
\alias{compare_all_models}
\title{compare_all_models}
\usage{
compare_all_models(model_list)
}
\arguments{
\item{model_list}{list of models}
}
\value{
comparison of models
}
\description{
compare_all_models
}
\examples{
data <- carData::Salaries
model_1 <- lm(salary ~ yrs.since.phd, data)
model_2 <- lm(salary ~ yrs.since.phd * rank, data)
model_3 <- lm(salary ~ yrs.since.phd * rank * sex, data)
compare_all_models(list(model_1, model_2, model_3))
}
