% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_brains.R
\name{compare_brains}
\alias{compare_brains}
\title{compare_brains}
\usage{
compare_brains(
  a,
  b,
  mask,
  ci = 0.95,
  legend_title = "Intensity",
  out_path = NULL,
  out_width = 14,
  out_height = 7,
  ...
)
}
\arguments{
\item{a}{first image to compare}

\item{b}{second image to compare}

\item{mask}{mask image}

\item{ci}{confidence interval for bland-altman plots (default: 0.95)}

\item{legend_title}{legend title (default: intensity)}

\item{out_path}{path to save image}

\item{out_width}{width of saved image in inches (default: 11)}

\item{out_height}{height of saved image in inches (default: 8.5)}

\item{...}{additional parameters for \code{plot_brain()}}
}
\value{
ggplot figure comparing two images
}
\description{
compare_brains
}
\seealso{
Other neuroimaging helpers:
\code{\link[=convet_lut_fs2bids]{convet_lut_fs2bids()}},
\code{\link[=extract_brain_slice]{extract_brain_slice()}},
\code{\link[=get_3dstats]{get_3dstats()}},
\code{\link[=plot_brain]{plot_brain()}},
\code{\link[=plot_brain_slice]{plot_brain_slice()}},
\code{\link[=split_dseg2mask]{split_dseg2mask()}}
}
\concept{neuroimaging}
\concept{neuroimaging helpers}
\concept{visualization}
