% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_models}
\alias{compare_models}
\title{compare_models}
\usage{
compare_models(model_1, model_2)
}
\arguments{
\item{model_1}{a model to compare}

\item{model_2}{the other model to compare}
}
\value{
comparison of the two models
}
\description{
compare_models
}
\examples{
data <- carData::Salaries
model_1 <- lm(salary ~ yrs.since.phd * rank, data)
model_2 <- lm(salary ~ yrs.since.phd * rank * sex, data)
compare_models(model_1, model_2)
}
