% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_long.R
\name{cor_long}
\alias{cor_long}
\title{cor_long}
\usage{
cor_long(data)
}
\arguments{
\item{data}{data to correlate and transform to long format}
}
\description{
cor_long
}
\examples{
cor_long(mtcars)
}
\concept{stats}
