% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cor_heatmap.R
\name{cor_plot}
\alias{cor_plot}
\title{cor_plot}
\usage{
cor_plot(data)
}
\arguments{
\item{data}{data to create a heat map from}
}
\value{
ggplot2 correlation heatmap.
}
\description{
cor_plot
}
\examples{
cor_plot(mtcars)
}
\seealso{
Other plotting helpers: 
\code{\link{move_legend_to_btm}()},
\code{\link{plot_age}()},
\code{\link{plot_bland_altman}()},
\code{\link{plot_cor_heatmap}()},
\code{\link{plot_cor}()},
\code{\link{plot_scree}()}
}
\concept{plotting helpers}
\concept{visualization}
