% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_brain_slice.R
\name{extract_brain_slice}
\alias{extract_brain_slice}
\title{extract_brain_slice}
\usage{
extract_brain_slice(data, dir, slice)
}
\arguments{
\item{data}{3D brain data}

\item{dir}{direction to extract slice (choices: i, j, or k)}

\item{slice}{integer from 1 to slice max}
}
\value{
data.frame of the slice in long format (columns: x, y, value)
}
\description{
extract_brain_slice
}
\examples{
brain <- array(1:27, dim = c(3, 3, 3))
extract_brain_slice(brain, dir = "k", slice = 2)
}
\seealso{
Other neuroimaging helpers: 
\code{\link{compare_brains}()},
\code{\link{convet_lut_fs2bids}()},
\code{\link{get_3dstats}()},
\code{\link{plot_brain}()},
\code{\link{plot_brain_slice}()},
\code{\link{split_dseg2mask}()}
}
\concept{neuroimaging}
\concept{neuroimaging helpers}
