% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_dimension.R
\name{flatten_dimension}
\alias{flatten_dimension}
\title{flatten_dimension}
\usage{
flatten_dimension(data)
}
\arguments{
\item{data}{multi-dimensional data to be flattened/reduced by a single dimension}
}
\value{
data that is flattened/reduced by a single dimension
}
\description{
flatten_dimension
}
\examples{
x <- array(1:24, dim = c(2, 3, 4))
flattened_once <- flatten_dimension(x)
str(flattened_once)
}
\seealso{
Other dimension flattening helpers: 
\code{\link{flatten_dimension_all}()}
}
\concept{data_wrangle}
\concept{dimension flattening helpers}
