% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_dimension_all.R
\name{flatten_dimension_all}
\alias{flatten_dimension_all}
\title{flatten_dimension_all}
\usage{
flatten_dimension_all(data)
}
\arguments{
\item{data}{multi-dimensional data to completely flattened/reduced to a single dimension}
}
\value{
data that is flattened to a single dimension
}
\description{
flatten_dimension_all
}
\examples{
x <- array(1:24, dim = c(2, 3, 4))
flatten_dimension_all(x)
}
\seealso{
Other dimension flattening helpers: 
\code{\link{flatten_dimension}()}
}
\concept{data_wrangle}
\concept{dimension flattening helpers}
