% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_3dstats.R
\name{get_3dstats}
\alias{get_3dstats}
\title{get_3dstats}
\usage{
get_3dstats(in_data, in_dseg, in_lut = NULL, out_path, overwrite = FALSE)
}
\arguments{
\item{in_data}{input path to data nifti file}

\item{in_dseg}{input path to discrete segmentation nifti file}

\item{in_lut}{input path to look up table}

\item{out_path}{output path to save stats tsv file}

\item{overwrite}{flag to overwrite (default: 0)}
}
\value{
A data.frame of regional 3D summary statistics; optionally writes it to \code{out_path}.
}
\description{
obtain summary metrics of data file from discrete segmentation. summary metrics include non-zero voxels, min, max, median, mean
}
\seealso{
Other neuroimaging helpers:
\code{\link[=compare_brains]{compare_brains()}},
\code{\link[=convet_lut_fs2bids]{convet_lut_fs2bids()}},
\code{\link[=extract_brain_slice]{extract_brain_slice()}},
\code{\link[=plot_brain]{plot_brain()}},
\code{\link[=plot_brain_slice]{plot_brain_slice()}},
\code{\link[=split_dseg2mask]{split_dseg2mask()}}
}
\concept{neuroimaging}
\concept{neuroimaging helpers}
