% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age_yrs.R
\name{get_age_yrs}
\alias{get_age_yrs}
\title{get_age_yrs}
\usage{
get_age_yrs(dob, ref_date = lubridate::today())
}
\arguments{
\item{dob}{date of birth}

\item{ref_date}{reference date (default: today's date)}
}
\description{
calculate age in years
}
\examples{
get_age_years('1776-07-04')
}
