% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_info.R
\name{info_msg}
\alias{info_msg}
\title{info_msg}
\usage{
info_msg(msg)
}
\arguments{
\item{msg}{info message to print}
}
\value{
Invisibly returns \code{NULL}; prints a formatted info message.
}
\description{
info_msg
}
\examples{
info_msg("example info")
}
\seealso{
Other console messaging helpers:
\code{\link[=error_msg]{error_msg()}},
\code{\link[=print_footer]{print_footer()}},
\code{\link[=print_header]{print_header()}},
\code{\link[=warning_msg]{warning_msg()}}
}
\concept{console messaging helpers}
