% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_optparse.R
\name{initialize_optparse}
\alias{initialize_optparse}
\title{initialize_optparse}
\usage{
initialize_optparse(func_name, func_path = NULL, out_path = NULL)
}
\arguments{
\item{func_name}{name of function [type: string]}

\item{func_path}{path to function (optional)}

\item{out_path}{path to write optparsed file (optional)}
}
\description{
creates an optparse template from a function and its arguments
}
\examples{
initialize_optparse("pmid2doi")
}
\concept{r_helper}
