% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_parallel_computing.R
\name{initialize_parallel_computing}
\alias{initialize_parallel_computing}
\title{initialize_parallel_computing}
\usage{
initialize_parallel_computing(save = 1)
}
\arguments{
\item{save}{number of cores to save/exclude from parallel computing (default: 1)}
}
\description{
initialize parallel computing for different types of parallel computing
}
