% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_scaled.R
\name{is_centered}
\alias{is_centered}
\title{is_centered}
\usage{
is_centered(x, eps = 1e-08)
}
\arguments{
\item{x}{a numeric vector or matrix}

\item{eps}{small error tolerance of being close enough to zero (default: 1e-8)}
}
\value{
vector of logicals
}
\description{
Checks if a numeric vector or matrix (column-wise) is mean-centered. In other words, checks if the mean is zero (within a specified tolerance)
}
\examples{
is_centered(scale(c(1, 2, 2, 3, 3, 3)))
}
\seealso{
Other check scale functions:
\code{\link[=is_scaled]{is_scaled()}},
\code{\link[=is_ss1]{is_ss1()}},
\code{\link[=is_z_scored]{is_z_scored()}}
}
\concept{check scale functions}
\concept{data_wrangle}
