% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_scaled.R
\name{is_scaled}
\alias{is_scaled}
\title{is_scaled}
\usage{
is_scaled(x, type = "z", eps = 1e-08)
}
\arguments{
\item{x}{a numeric vector or matrix}

\item{type}{type of scaling to check for. Options are: "mean_centered" (or "centered"), "z_scored" (or "z"), and "ss1". Default is "z".}

\item{eps}{small error tolerance of being close enough to zero (default: 1e-8)}
}
\value{
vector of logicals
}
\description{
Checks if a numeric vector or matrix (column-wise) is scaled according to a specified type
}
\examples{
is_scaled(scale(c(1, 2, 2, 3, 3, 3)))
}
\seealso{
Other check scale functions:
\code{\link[=is_centered]{is_centered()}},
\code{\link[=is_ss1]{is_ss1()}},
\code{\link[=is_z_scored]{is_z_scored()}}
}
\concept{check scale functions}
\concept{data_wrangle}
