% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_scaled.R
\name{is_ss1}
\alias{is_ss1}
\title{is_ss1}
\usage{
is_ss1(x, eps = 1e-08)
}
\arguments{
\item{x}{a numeric vector or matrix}

\item{eps}{small error tolerance of being close enough to zero (default: 1e-8)}
}
\value{
vector of logicals
}
\description{
Checks if a numeric vector or matrix (column-wise) has a sum of squares equal to 1 (within a specified tolerance)
}
\examples{
x <- c(1, 2, 2, 3, 3, 3)
is_ss1(scale(x) / sqrt(length(x)))
}
\seealso{
Other check scale functions: 
\code{\link{is_centered}()},
\code{\link{is_scaled}()},
\code{\link{is_z_scored}()}
}
\concept{check scale functions}
\concept{data_wrangle}
