% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_scaled.R
\name{is_z_scored}
\alias{is_z_scored}
\title{is_z_scored}
\usage{
is_z_scored(x, eps = 1e-08)
}
\arguments{
\item{x}{a numeric vector or matrix}

\item{eps}{small error tolerance of being close enough to zero (default: 1e-8)}
}
\value{
vector of logicals
}
\description{
Checks if a numeric vector or matrix (column-wise) is z-scored. In other words, checks if the mean is zero and the standard deviation is one (within a specified tolerance)
}
\examples{
is_z_scored(scale(c(1, 2, 2, 3, 3, 3)))
}
\seealso{
Other check scale functions: 
\code{\link{is_centered}()},
\code{\link{is_scaled}()},
\code{\link{is_ss1}()}
}
\concept{check scale functions}
\concept{data_wrangle}
