% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm2apa.R
\name{lm2apa}
\alias{lm2apa}
\title{lm2apa}
\usage{
lm2apa(m, terms = "all", level = 0.95, digits = 3, format = "html")
}
\arguments{
\item{m}{model from `lm()`}

\item{terms}{specify terms to print in APA format (default: all)}

\item{level}{specify confidence level (default: 0.95)}

\item{digits}{specify digit to round statistics (default: 3)}

\item{format}{specify format (default: html)}
}
\description{
converts `lm` model to APA format string
}
