% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer2apa.R
\name{lmer2apa}
\alias{lmer2apa}
\title{lmer2apa}
\usage{
lmer2apa(m, terms = "all", level = 0.95, digits = 3, format = "html")
}
\arguments{
\item{m}{model from \code{lmer()}}

\item{terms}{specify terms to print in APA format (default: all)}

\item{level}{specify confidence level (default: 0.95)}

\item{digits}{specify digit to round statistics (default: 3)}

\item{format}{specify format (default: html) (options: html, plain)}
}
\description{
converts \code{lmer()} model to APA format string
}
\examples{
lmer2apa(lmerTest::lmer(Reaction ~ 1 + Days + (1 + Days | Subject), lme4::sleepstudy), format = 'plain')
}
\seealso{
Other model2apa: 
\code{\link{cor2apa}()},
\code{\link{lm2apa}()},
\code{\link{model2apa}()}
}
\concept{model2apa}
\concept{stats}
