% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_terms_to_data.R
\name{map_terms_to_data}
\alias{map_terms_to_data}
\title{Map Model Terms to Raw Data Variables}
\usage{
map_terms_to_data(model, terms)
}
\arguments{
\item{model}{A fitted model object (lm, glm, lmer, lmerTest, etc.).}

\item{terms}{Character vector of one or more term/variable names.}
}
\value{
A character vector of corresponding raw variable names from the model frame.
}
\description{
Map Model Terms to Raw Data Variables
}
\concept{stats}
