% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2apa.R
\name{model2apa}
\alias{model2apa}
\title{model2apa}
\usage{
model2apa(m, terms = "all", level = 0.95, digits = 3, format = "html")
}
\arguments{
\item{m}{model from \code{lmer()}}

\item{terms}{specify terms to print in APA format (default: all)}

\item{level}{specify confidence level (default: 0.95)}

\item{digits}{specify digit to round statistics (default: 3)}

\item{format}{specify format (default: html) (options: html, plain)}
}
\description{
converts a model to APA format string
}
\examples{
# cor.test
model2apa(cor.test(carData::Salaries$yrs.since.phd, carData::Salaries$salary), format = 'plain')

# lm
model2apa(lm(mpg ~ wt, data = mtcars), format = 'plain')

# lme
model2apa(lmerTest::lmer(Reaction ~ 1 + Days + (1 + Days | Subject), lme4::sleepstudy), format = 'plain')
}
\seealso{
Other model2apa: 
\code{\link{cor2apa}()},
\code{\link{lm2apa}()},
\code{\link{lmer2apa}()}
}
\concept{model2apa}
\concept{stats}
