% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_binary_lasso.R
\name{model_binary_lasso}
\alias{model_binary_lasso}
\title{model_binary_lasso}
\usage{
model_binary_lasso(data, y)
}
\arguments{
\item{data}{data to be analyzed}

\item{y}{name/string of outcome to be predicted that is within the data}
}
\value{
lasso results
}
\description{
model_binary_lasso
}
\examples{
# to be added
}
