% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_binary_ridge.R
\name{model_binary_ridge}
\alias{model_binary_ridge}
\title{model_binary_ridge}
\usage{
model_binary_ridge(data, y)
}
\arguments{
\item{data}{clean data.frame of all variables to be used}

\item{y}{name of binary outcome within the data.frame to be predicted}
}
\value{
ridge model results
}
\description{
predict binary outcome using ridge regression
}
\examples{
# to be added
}
