% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_categorical_all.R
\name{model_categorical_all}
\alias{model_categorical_all}
\title{model_categorical_all}
\usage{
model_categorical_all(x, y)
}
\arguments{
\item{x}{features used in prediction}

\item{y}{outcome to predict}
}
\value{
A tibble with model names and fitted categorical models.
}
\description{
model_categorical_all
}
\examples{
\dontrun{
x <- iris[, c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")]
y <- iris$Species
model_categorical_all(x, y)
}
}
\seealso{
Other categorical model wrappers: 
\code{\link{model_categorical_elastic_net}()},
\code{\link{model_categorical_glm}()},
\code{\link{model_categorical_lasso}()},
\code{\link{model_categorical_ridge}()},
\code{\link{model_categorical_svm}()}
}
\concept{categorical model wrappers}
\concept{model_uber}
