% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_categorical_elastic_net.R
\name{model_categorical_elastic_net}
\alias{model_categorical_elastic_net}
\title{model_categorical_elastic_net}
\usage{
model_categorical_elastic_net(x, y, alpha_list = seq(1e-04, 0.9999, 1e-04))
}
\arguments{
\item{x}{features/predictors used in prediction of outcome}

\item{y}{categorical outcome to predict}

\item{alpha_list}{vector of alpha values to cross-validate}
}
\value{
elastic net results
}
\description{
model_categorical_elastic_net
}
\examples{
\dontrun{
x <- iris[, c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")]
y <- iris$Species
model_categorical_elastic_net(x, y, alpha_list = c(0.25, 0.5, 0.75))
}
}
\seealso{
Other categorical model wrappers:
\code{\link[=model_categorical_all]{model_categorical_all()}},
\code{\link[=model_categorical_glm]{model_categorical_glm()}},
\code{\link[=model_categorical_lasso]{model_categorical_lasso()}},
\code{\link[=model_categorical_ridge]{model_categorical_ridge()}},
\code{\link[=model_categorical_svm]{model_categorical_svm()}}
}
\concept{categorical model wrappers}
\concept{model_categorical}
