% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_categorical_glm.R
\name{model_categorical_glm}
\alias{model_categorical_glm}
\title{model_categorical_glm}
\usage{
model_categorical_glm(x, y)
}
\arguments{
\item{x}{features/predictors used in prediction of outcome}

\item{y}{categorical outcome to predict}
}
\value{
Fitted binomial GLM or multinomial model.
}
\description{
model_categorical_glm
}
\examples{
x <- iris[, c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")]
y <- iris$Species
model_categorical_glm(x, y)
}
\seealso{
Other categorical model wrappers:
\code{\link[=model_categorical_all]{model_categorical_all()}},
\code{\link[=model_categorical_elastic_net]{model_categorical_elastic_net()}},
\code{\link[=model_categorical_lasso]{model_categorical_lasso()}},
\code{\link[=model_categorical_ridge]{model_categorical_ridge()}},
\code{\link[=model_categorical_svm]{model_categorical_svm()}}
}
\concept{categorical model wrappers}
\concept{model_categorical}
