% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_categorical_ridge.R
\name{model_categorical_ridge}
\alias{model_categorical_ridge}
\title{model_categorical_ridge}
\usage{
model_categorical_ridge(x, y)
}
\arguments{
\item{x}{features/predictors used in prediction of outcome}

\item{y}{categorical outcome to predict}
}
\value{
Fitted ridge regression model from \code{glmnet}.
}
\description{
model_categorical_ridge
}
\examples{
x <- iris[, c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")]
y <- iris$Species
model_categorical_ridge(x, y)
}
\seealso{
Other categorical model wrappers:
\code{\link[=model_categorical_all]{model_categorical_all()}},
\code{\link[=model_categorical_elastic_net]{model_categorical_elastic_net()}},
\code{\link[=model_categorical_glm]{model_categorical_glm()}},
\code{\link[=model_categorical_lasso]{model_categorical_lasso()}},
\code{\link[=model_categorical_svm]{model_categorical_svm()}}
}
\concept{categorical model wrappers}
\concept{model_categorical}
