% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_continuous_elastic_net.R
\name{model_continuous_elastic_net}
\alias{model_continuous_elastic_net}
\title{model_continuous_elastic_net}
\usage{
model_continuous_elastic_net(data, y, alpha_list = seq(1e-04, 0.9999, 1e-04))
}
\arguments{
\item{data}{data to be analyzed}

\item{y}{name/string of outcome to be predicted that is within the data}

\item{alpha_list}{vector of alpha to cross-validate (default: seq(0.0001, 0.9999, 0.0001))}
}
\value{
elastic net results
}
\description{
model_continuous_elastic_net
}
\examples{
\dontrun{
data <- data.frame(y = mtcars$mpg, wt = mtcars$wt, hp = mtcars$hp)
model_continuous_elastic_net(data, "y", alpha_list = c(0.25, 0.5, 0.75))
}
}
\seealso{
Other continuous model wrappers:
\code{\link[=model_continuous_lasso]{model_continuous_lasso()}},
\code{\link[=model_continuous_ridge]{model_continuous_ridge()}},
\code{\link[=model_continuous_svm_linear]{model_continuous_svm_linear()}}
}
\concept{continuous model wrappers}
\concept{model_continuous}
