% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_continuous_ridge.R
\name{model_continuous_ridge}
\alias{model_continuous_ridge}
\title{model_continuous_ridge}
\usage{
model_continuous_ridge(data, y)
}
\arguments{
\item{data}{data to be analyzed}

\item{y}{name/string of outcome to be predicted that is within the data}
}
\value{
ridge results
}
\description{
model_continuous_ridge
}
\examples{
data <- data.frame(y = mtcars$mpg, wt = mtcars$wt, hp = mtcars$hp)
model_continuous_ridge(data, "y")
}
\seealso{
Other continuous model wrappers: 
\code{\link{model_continuous_elastic_net}()},
\code{\link{model_continuous_lasso}()},
\code{\link{model_continuous_svm_linear}()}
}
\concept{continuous model wrappers}
\concept{model_continuous}
