% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_continuous_svm_linear.R
\name{model_continuous_svm_linear}
\alias{model_continuous_svm_linear}
\title{model_continuous_svm_linear}
\usage{
model_continuous_svm_linear(data, y)
}
\arguments{
\item{data}{clean data.frame of predictors and outcome to be analyzed}

\item{y}{continuous outcome name within the data.frame}
}
\value{
linear svm model
}
\description{
model continuous linear support vector machine
}
\examples{
data <- data.frame(y = mtcars$mpg, wt = mtcars$wt, hp = mtcars$hp)
model_continuous_svm_linear(data, "y")
}
\seealso{
Other continuous model wrappers:
\code{\link[=model_continuous_elastic_net]{model_continuous_elastic_net()}},
\code{\link[=model_continuous_lasso]{model_continuous_lasso()}},
\code{\link[=model_continuous_ridge]{model_continuous_ridge()}}
}
\concept{continuous model wrappers}
\concept{model_continuous}
