% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_legend_to_btm.R
\name{move_legend_to_btm}
\alias{move_legend_to_btm}
\title{move_legend_to_btm}
\usage{
move_legend_to_btm(learn = FALSE)
}
\arguments{
\item{learn}{logical, whether to print the equivalent ggplot2 theme code}
}
\value{
ggplot2 theme object
}
\description{
move_legend_to_btm
}
\examples{
plot_cor("wt", "mpg", mtcars) + move_legend_to_btm()
}
\seealso{
Other plotting helpers:
\code{\link[=cor_plot]{cor_plot()}},
\code{\link[=plot_age]{plot_age()}},
\code{\link[=plot_bland_altman]{plot_bland_altman()}},
\code{\link[=plot_cor]{plot_cor()}},
\code{\link[=plot_cor_heatmap]{plot_cor_heatmap()}},
\code{\link[=plot_scree]{plot_scree()}}
}
\concept{plotting helpers}
\concept{visualization}
