% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_table_html.R
\name{nice_table_html}
\alias{nice_table_html}
\title{nice_table_html}
\usage{
nice_table_html(table)
}
\arguments{
\item{table}{a data.frame/table to pass through kable}
}
\value{
a scrollable HTML table
}
\description{
nice_table_html
}
\examples{
nice_table_html(head(mtcars))
}
\seealso{
Other table helpers:
\code{\link[=nice_table]{nice_table()}},
\code{\link[=nice_table_html_short]{nice_table_html_short()}}
}
\concept{table helpers}
