% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_table_html_short.R
\name{nice_table_html_short}
\alias{nice_table_html_short}
\title{nice_table_html_short}
\usage{
nice_table_html_short(table, height = "200px")
}
\arguments{
\item{table}{a data.frame/table to pass through kable}

\item{height}{height of the scroll box}
}
\value{
a short scrollable HTML table
}
\description{
nice_table_html_short
}
\examples{
nice_table_html_short(mtcars, height = "150px")
}
\seealso{
Other table helpers: 
\code{\link{nice_table_html}()},
\code{\link{nice_table}()}
}
\concept{table helpers}
