% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bland_altman.R
\name{plot_bland_altman}
\alias{plot_bland_altman}
\title{plot_bland_altman}
\usage{
plot_bland_altman(a, b, data, ci = 0.95, txt_offset = 0.25)
}
\arguments{
\item{a}{first variable in comparison}

\item{b}{second variable in comparison}

\item{data}{data of the variables}

\item{ci}{confidence interval (default: 0.95)}

\item{txt_offset}{text offset for labels (default: 0.25)}
}
\description{
plot_bland_altman
}
\examples{
# note: not the best example, since it should be conceptually similar metrics
plot_bland_altman('yrs.since.phd', 'yrs.service', carData::Salaries)
}
\concept{visualization}
