% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_brain.R
\name{plot_brain}
\alias{plot_brain}
\title{plot_brain}
\usage{
plot_brain(data, indices = NULL, legend_position = NULL, ...)
}
\arguments{
\item{data}{3D brain data}

\item{indices}{slices index of i, j, and k to plot}

\item{legend_position}{position of the legend can be bottom or NULL (default)}

\item{...}{extra options to pass to `plot_brain_slice`}
}
\value{
ggplot2 figure of all three directions (i, j, k)
}
\description{
plot_brain
}
\concept{neuroimaging}
\concept{visualization}
