% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_brain_slice.R
\name{plot_brain_slice}
\alias{plot_brain_slice}
\title{plot_brain_slice}
\usage{
plot_brain_slice(
  df,
  title = NULL,
  fill_limits = "min_max",
  fill_low = "black",
  fill_mid = NULL,
  fill_midpoint = 0,
  fill_high = "white",
  legend_position = NULL,
  legend_title = NULL,
  slice_label_offset = 0.25
)
}
\arguments{
\item{df}{brain slice data (output of `plot_brain_slice`)}

\item{title}{title of plot (default: NULL)}

\item{fill_limits}{limits of the fill}

\item{fill_low}{color of the lowest limit (default: black)}

\item{fill_mid}{color of the midpoint. assigning a color automatically using `scale_fill_gradient2` rather `scale_fill_gradient`}

\item{fill_midpoint}{midpoint value (default: 0)}

\item{fill_high}{color of the highest limit (default: white)}

\item{legend_position}{position of the legend (default: NULL). can also be 'bottom'}

\item{legend_title}{title of fill legend (default: NULL)}

\item{slice_label_offset}{slice label offset (default: 0.25)}
}
\value{

}
\description{
plot_brain_slice
}
\concept{neuroimaging}
\concept{visualization}
