% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cor.R
\name{plot_cor}
\alias{plot_cor}
\title{plot_cor}
\usage{
plot_cor(x, y, data)
}
\arguments{
\item{x}{x variable}

\item{y}{y variable}

\item{data}{data that contains the x and y variables}
}
\value{
ggplot2 figure with correlation statistics
}
\description{
plot_cor
}
\examples{
plot_cor('yrs.service', 'yrs.since.phd', carData::Salaries)
}
\seealso{
Other plotting helpers:
\code{\link[=cor_plot]{cor_plot()}},
\code{\link[=move_legend_to_btm]{move_legend_to_btm()}},
\code{\link[=plot_age]{plot_age()}},
\code{\link[=plot_bland_altman]{plot_bland_altman()}},
\code{\link[=plot_cor_heatmap]{plot_cor_heatmap()}},
\code{\link[=plot_scree]{plot_scree()}}
}
\concept{plotting helpers}
\concept{visualization}
