% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cor_heatmap.R
\name{plot_cor_heatmap}
\alias{plot_cor_heatmap}
\title{plot_cor_heatmap}
\usage{
plot_cor_heatmap(data)
}
\arguments{
\item{data}{data to create a heat map from}
}
\value{
ggplot figure
}
\description{
plot_cor_heatmap
}
\examples{
plot_cor_heatmap(mtcars)
}
\seealso{
Other plotting helpers:
\code{\link[=cor_plot]{cor_plot()}},
\code{\link[=move_legend_to_btm]{move_legend_to_btm()}},
\code{\link[=plot_age]{plot_age()}},
\code{\link[=plot_bland_altman]{plot_bland_altman()}},
\code{\link[=plot_cor]{plot_cor()}},
\code{\link[=plot_scree]{plot_scree()}}
}
\concept{plotting helpers}
\concept{visualization}
