% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_steps.R
\name{plot_steps}
\alias{plot_steps}
\title{plot_steps}
\usage{
plot_steps(
  data,
  out_path = NULL,
  subjid = "sub",
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{data}{path to input csv file of subjects by rows (character) OR a data.frame containing the data (required)}

\item{out_path}{path to save output image (optional)}

\item{subjid}{columns that identify the rows. if more than one, will combine into key1-value1_key2-value2_...keyN-valueN}

\item{title}{title of plot}

\item{subtitle}{subtitle of plot}
}
\value{
ggplot2 heatmap of steps
}
\description{
plot_steps
}
\examples{
df <- data.frame(
  sub = c("sub01", "sub02", "sub03"),
  `1. Step 1` = c(1, 1, 1),
  `2. Step 2` = c(1, 1, 0),
  `3. Step 3` = c(1, 0, 0),
  check.names = FALSE
)
plot_steps(df)
}
\concept{visualization}
