% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmid2funcs.R
\name{pmid2abstract}
\alias{pmid2abstract}
\title{pmid2abstract}
\usage{
pmid2abstract(pmid)
}
\arguments{
\item{pmid}{PubMed identifier}
}
\value{
Character abstract text for the PMID, when available.
}
\description{
pmid2abstract
}
\examples{
\dontrun{
pmid2abstract(33317393)
}
}
\seealso{
Other PubMed helpers: 
\code{\link{pmcid2pdf}()},
\code{\link{pmid2bibtex}()},
\code{\link{pmid2doi}()},
\code{\link{pmid2pmcid}()}
}
\concept{PubMed helpers}
\concept{references_pubmed}
