% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmid2funcs.R
\name{pmid2bibtex}
\alias{pmid2bibtex}
\title{pmid2bibtex}
\usage{
pmid2bibtex(pmid, key_as_first_author_year = TRUE)
}
\arguments{
\item{pmid}{PubMed identifier}

\item{key_as_first_author_year}{logical, whether to use first-author-year as the BibTeX key}
}
\value{
Character string containing BibTeX citation text.
}
\description{
pmid2bibtex
}
\examples{
\dontrun{
pmid2bibtex(33317393)
}
}
\seealso{
Other PubMed helpers: 
\code{\link{pmcid2pdf}()},
\code{\link{pmid2abstract}()},
\code{\link{pmid2doi}()},
\code{\link{pmid2pmcid}()}
}
\concept{PubMed helpers}
\concept{references_pubmed}
