% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmid2funcs.R
\name{pmid2pmcid}
\alias{pmid2pmcid}
\title{pmid2pmcid}
\usage{
pmid2pmcid(pmid)
}
\arguments{
\item{pmid}{PubMed identifier}
}
\value{
Character PMCID for the PMID, when available.
}
\description{
pmid2pmcid
}
\examples{
\dontrun{
pmid2pmcid(33317393)
}
}
\seealso{
Other PubMed helpers:
\code{\link[=pmcid2pdf]{pmcid2pdf()}},
\code{\link[=pmid2abstract]{pmid2abstract()}},
\code{\link[=pmid2bibtex]{pmid2bibtex()}},
\code{\link[=pmid2doi]{pmid2doi()}}
}
\concept{PubMed helpers}
\concept{references_pubmed}
