% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_info.R
\name{print_footer}
\alias{print_footer}
\title{print_footer}
\usage{
print_footer(start_datetime)
}
\arguments{
\item{start_datetime}{start time from `print_header()`}
}
\value{
Invisibly returns \code{NULL}; prints elapsed time.
}
\description{
print_footer
}
\seealso{
Other console messaging helpers: 
\code{\link{error_msg}()},
\code{\link{info_msg}()},
\code{\link{print_header}()},
\code{\link{warning_msg}()}
}
\concept{console messaging helpers}
