% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_info.R
\name{print_header}
\alias{print_header}
\title{print_header}
\usage{
print_header()
}
\value{
time
}
\description{
print_header
}
\examples{
start_time <- print_header()
}
\seealso{
Other console messaging helpers: 
\code{\link{error_msg}()},
\code{\link{info_msg}()},
\code{\link{print_footer}()},
\code{\link{warning_msg}()}
}
\concept{console messaging helpers}
