% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_sq_to_adj_r_sq.R
\name{r_sq_to_adj_r_sq}
\alias{r_sq_to_adj_r_sq}
\title{r_sq_to_adj_r_sq}
\usage{
r_sq_to_adj_r_sq(r_sq, n, n_p)
}
\arguments{
\item{r_sq}{r-squared ($R^2$) value}

\item{n}{number of subjects/rows}

\item{n_p}{number of parameters}
}
\value{
adjusted r-squared ($R^2_{Adj.}$) value
}
\description{
Convert r-squared ($R^2$) to adjusted r-squared ($R^2_{Adj.}$).
}
\examples{
r_sq_to_adj_r_sq(.5, 100, 5)
}
\seealso{
Other model summary helpers:
\code{\link[=get_lm_multicollinearity]{get_lm_multicollinearity()}},
\code{\link[=models2coefs]{models2coefs()}},
\code{\link[=models2omni]{models2omni()}},
\code{\link[=renamed_tidy]{renamed_tidy()}},
\code{\link[=tidy_es]{tidy_es()}},
\code{\link[=tidy_es_cope_F]{tidy_es_cope_F()}},
\code{\link[=tidy_es_cope_t]{tidy_es_cope_t()}},
\code{\link[=tidy_es_lm]{tidy_es_lm()}},
\code{\link[=tidy_es_lmer]{tidy_es_lmer()}}
}
\concept{model summary helpers}
\concept{stats}
