% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_colnames.R
\name{rename_colnames}
\alias{rename_colnames}
\title{rename_colnames}
\usage{
rename_colnames(data, new_colnames)
}
\arguments{
\item{data}{data.frame to be renamed}

\item{new_colnames}{a vector of new column names}
}
\value{
data.frame with the new column names
}
\description{
rename column names
}
\examples{
rename_colnames(
  data.frame(a = 1:3, b = 4:6),
  c("participant_id", "score")
)
}
\seealso{
Other column name helpers: 
\code{\link{append_colnames}()},
\code{\link{append_prefix_to_colnames}()},
\code{\link{append_suffix_to_colnames}()},
\code{\link{tidy_colnames}()}
}
\concept{column name helpers}
