% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed_tidy.R
\name{renamed_tidy}
\alias{renamed_tidy}
\title{renamed_tidy}
\usage{
renamed_tidy(model)
}
\arguments{
\item{model}{model fit from \code{lm()}}
}
\value{
data.frame
}
\description{
Minor modification of the \code{tidy()} function that renames the columns to be more intuitive and adds the left-hand and operator columns for lm objects
}
\examples{
renamed_tidy(lm(mpg ~ wt, data = mtcars))
}
\seealso{
Other model summary helpers:
\code{\link[=get_lm_multicollinearity]{get_lm_multicollinearity()}},
\code{\link[=models2coefs]{models2coefs()}},
\code{\link[=models2omni]{models2omni()}},
\code{\link[=r_sq_to_adj_r_sq]{r_sq_to_adj_r_sq()}},
\code{\link[=tidy_es]{tidy_es()}},
\code{\link[=tidy_es_cope_F]{tidy_es_cope_F()}},
\code{\link[=tidy_es_cope_t]{tidy_es_cope_t()}},
\code{\link[=tidy_es_lm]{tidy_es_lm()}},
\code{\link[=tidy_es_lmer]{tidy_es_lmer()}}
}
\concept{model summary helpers}
\concept{stats}
