% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_to_write.R
\name{safe_to_write}
\alias{safe_to_write}
\title{safe_to_write}
\usage{
safe_to_write(f, overwrite = FALSE)
}
\arguments{
\item{f}{path to file}

\item{overwrite}{logical, whether to overwrite if file already exists (default: FALSE)}
}
\value{
logical, whether it is safe to write to the file
}
\description{
safe_to_write
}
\examples{
safe_to_write(file.path(tempdir(), "new-file.txt"))
}
\seealso{
Other path helpers: 
\code{\link{check_in_paths}()},
\code{\link{check_out_paths}()},
\code{\link{eval_cmd}()},
\code{\link{mkdir_if_dne}()},
\code{\link{stop_if_dne}()},
\code{\link{stop_if_e}()}
}
\concept{path helpers}
