% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dseg.R
\name{split_dseg}
\alias{split_dseg}
\title{split_dseg}
\usage{
split_dseg(in_dseg, in_lut, out_dir, overwrite = FALSE)
}
\arguments{
\item{in_dseg}{path to discrete segmentation file}

\item{in_lut}{path to look up table .tsv file (bids format)}

\item{out_dir}{output directory for the individual segmentation files}

\item{overwrite}{flag to overwrite output files (default: 0)}
}
\description{
split_dseg
}
\concept{neuroimaging}
