% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_if_dne.R
\name{stop_if_dne}
\alias{stop_if_dne}
\title{stop_if_dne}
\usage{
stop_if_dne(file)
}
\arguments{
\item{file}{file path to check}
}
\value{
Invisibly returns \code{NULL}; errors if \code{file} does not exist.
}
\description{
stop_if_dne
}
\examples{
stop_if_dne(tempdir())
}
\seealso{
Other path helpers:
\code{\link[=check_in_paths]{check_in_paths()}},
\code{\link[=check_out_paths]{check_out_paths()}},
\code{\link[=eval_cmd]{eval_cmd()}},
\code{\link[=mkdir_if_dne]{mkdir_if_dne()}},
\code{\link[=safe_to_write]{safe_to_write()}},
\code{\link[=stop_if_e]{stop_if_e()}}
}
\concept{path helpers}
