% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_if_e.R
\name{stop_if_e}
\alias{stop_if_e}
\title{stop_if_e}
\usage{
stop_if_e(file)
}
\arguments{
\item{file}{file path to check}
}
\value{
Invisibly returns \code{NULL}; errors if \code{file} exists.
}
\description{
stop_if_e
}
\examples{
stop_if_e(file.path(tempdir(), "missing-file.txt"))
}
\seealso{
Other path helpers:
\code{\link[=check_in_paths]{check_in_paths()}},
\code{\link[=check_out_paths]{check_out_paths()}},
\code{\link[=eval_cmd]{eval_cmd()}},
\code{\link[=mkdir_if_dne]{mkdir_if_dne()}},
\code{\link[=safe_to_write]{safe_to_write()}},
\code{\link[=stop_if_dne]{stop_if_dne()}}
}
\concept{path helpers}
