% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_colnames.R
\name{tidy_colnames}
\alias{tidy_colnames}
\title{tidy_colnames}
\usage{
tidy_colnames(data)
}
\arguments{
\item{data}{data.frame that contains the column names to be converted}
}
\value{
data.frame with tidy column names
}
\description{
makes the data.frame have column names that are in tidy format.
}
\details{
currently, does the following modifications:
1. lower-case
2. replaces:
   - periods with underscores
   - space with underscores
   - double underscore with one underscore
   - parentheses with underscore
   - brackets with underscore
   - forward and backward slashes with underscore
   - dash with underscore
   - e_mail with email
   - removes trailing white space and trailing underscores
}
\examples{
# before
colnames(carData::Salaries)

# after
df <- carData::Salaries \%>\%
tidy_colnames(.)
colnames(df)
}
