% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_es.R
\name{tidy_es_cope_t}
\alias{tidy_es_cope_t}
\title{tidy_es_cope_t}
\usage{
tidy_es_cope_t(
  model,
  ci = 0.95,
  mcc = c("single-step", "none", "uncorrected", "fdr", "FDR", "tukey", "tukey's HSD",
    "Tukey")
)
}
\arguments{
\item{model}{contrast model from \code{multcomp::glht()}}

\item{ci}{confidence interval (0, 1)}

\item{mcc}{multiple comparison correction method. Options include "single-step" (default, which is what glht provides for Tukey's HSD), "none" (uncorrected), "fdr", "FDR", "tukey", "tukey's HSD", and "Tukey".}
}
\value{
data.frame
}
\description{
a modification of the tidy command that adds the effect
size (confidence intervals for the contrast coefficient and adjusted R^2)
for linear hypothesis testing contrasts (t-test) using glht
}
\examples{
model_fit <- lm(salary ~ rank + discipline, data = carData::Salaries)
c <- gen_contrast_ss(model_fit, x = "rank")[-1, ]
model_fit_cope_t <- multcomp::glht(model_fit, c)
tidy_es_cope_t(model_fit_cope_t)
}
\seealso{
Other model summary helpers:
\code{\link[=get_lm_multicollinearity]{get_lm_multicollinearity()}},
\code{\link[=models2coefs]{models2coefs()}},
\code{\link[=models2omni]{models2omni()}},
\code{\link[=r_sq_to_adj_r_sq]{r_sq_to_adj_r_sq()}},
\code{\link[=renamed_tidy]{renamed_tidy()}},
\code{\link[=tidy_es]{tidy_es()}},
\code{\link[=tidy_es_cope_F]{tidy_es_cope_F()}},
\code{\link[=tidy_es_lm]{tidy_es_lm()}},
\code{\link[=tidy_es_lmer]{tidy_es_lmer()}}

Other contrast or COPE helpers:
\code{\link[=check_contrast_orthogonality]{check_contrast_orthogonality()}},
\code{\link[=gen_contrast_blank]{gen_contrast_blank()}},
\code{\link[=gen_contrast_ss]{gen_contrast_ss()}},
\code{\link[=gen_data_jn]{gen_data_jn()}},
\code{\link[=plot.jn_df]{plot.jn_df()}},
\code{\link[=tidy_es_cope_F]{tidy_es_cope_F()}}
}
\concept{contrast or COPE helpers}
\concept{model summary helpers}
\concept{stats}
