% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryCatch2.R
\name{tryCatch2}
\alias{tryCatch2}
\title{tryCatch2}
\usage{
tryCatch2(expr)
}
\arguments{
\item{expr}{expression to evaluate with error printing}
}
\value{
result of \code{expr}, or printed error if \code{expr} fails
}
\description{
tryCatch2
}
\examples{
tryCatch2(1 + 1)
tryCatch2(stop("example error"))
}
\concept{error helpers}
\concept{r_helper}
