% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_info.R
\name{warning_msg}
\alias{warning_msg}
\title{warning_msg}
\usage{
warning_msg(msg)
}
\arguments{
\item{msg}{warning message to print}
}
\value{
Invisibly returns \code{NULL}; prints a formatted warning message.
}
\description{
warning_msg
}
\examples{
warning_msg("example warning")
}
\seealso{
Other console messaging helpers: 
\code{\link{error_msg}()},
\code{\link{info_msg}()},
\code{\link{print_footer}()},
\code{\link{print_header}()}
}
\concept{console messaging helpers}
