% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_colnames.R
\name{append_colnames}
\alias{append_colnames}
\title{append_colnames}
\usage{
append_colnames(data, prefix = NULL, suffix = NULL)
}
\arguments{
\item{data}{data.frame that contains the column names to change with a prefix}

\item{prefix}{string to add to the beginning of all of the column names of the data.frame}

\item{suffix}{string to add to the end of all of the column names of the data.frame}
}
\value{
data.frame with new column names with prefix string appended to the column names
}
\description{
add a string to beginning and/or ending of all the column names
}
\examples{
df <- append_colnames(carData::Salaries, prefix = 'p_', suffix = '_s')
head(df)
}
\concept{data_wrangle}
