% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_suffix_to_colnames.R
\name{append_suffix_to_colnames}
\alias{append_suffix_to_colnames}
\title{append_suffix_to_colnames}
\usage{
append_suffix_to_colnames(data, suffix)
}
\arguments{
\item{data}{data.frame that contains the column names to change with a suffix}

\item{suffix}{string to add to the end of all of the column names of the data.frame}
}
\value{
data.frame with new column names with suffix string appended to the column names
}
\description{
add a string to end of all the column names
}
\examples{
# to be added
}
