% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_suffix_to_colnames.R
\name{append_suffix_to_colnames}
\alias{append_suffix_to_colnames}
\title{append_suffix_to_colnames}
\usage{
append_suffix_to_colnames(data, suffix)
}
\arguments{
\item{suffix}{}
}
\description{
append_suffix_to_colnames
}
