% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_contrast_orthogonality.R
\name{check_contrast_orthogonality}
\alias{check_contrast_orthogonality}
\title{check_contrast_orthogonality}
\usage{
check_contrast_orthogonality(x)
}
\arguments{
\item{x}{contrast table/matrix}
}
\value{
logical (TRUE if both rules are met and FALSE if not)
}
\description{
checks to make sure that a contrast of a factor is orthogonal.
}
\details{
checks to make sure that a contrast of a factor is orthogonal by ensuring
(1) sum of each contrast equals 0
and (2) sum of each product of each contrast pair equals 0
}
\examples{
contrast_example <- data.frame(c1 = c(1/3, 1/3, -2/3),
                               c2 = c(-1/2, 1/2, 0))
check_contrast_orthogonality(contrast_example)
}
\seealso{
Other statistical helpers:
\code{\link[=calculate_sdt]{calculate_sdt()}}

Other contrast or COPE helpers:
\code{\link[=gen_contrast_blank]{gen_contrast_blank()}},
\code{\link[=gen_contrast_ss]{gen_contrast_ss()}},
\code{\link[=gen_data_jn]{gen_data_jn()}},
\code{\link[=plot.jn_df]{plot.jn_df()}},
\code{\link[=tidy_es_cope_F]{tidy_es_cope_F()}},
\code{\link[=tidy_es_cope_t]{tidy_es_cope_t()}}
}
\concept{contrast or COPE helpers}
\concept{statistical helpers}
\concept{stats}
