% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_contrast_orthogonality.R
\name{check_contrast_orthogonality}
\alias{check_contrast_orthogonality}
\title{check_contrast_orthogonality}
\usage{
check_contrast_orthogonality(x)
}
\arguments{
\item{x}{contrast table/matrix}
}
\description{
checks to make sure that a contrast of a factor is orthogonal.
}
\details{
checks to make sure that a contrast of a factor is orthogonal by ensuring
(1) sum of each contrast equals 0
and (2) sum of each product of each contrast pair equals 0
}
\examples{
contrast_example <- data.frame(c1 = c(1/3, 1/3, -2/3),
                               c2 = c(-1/2, 1/2, 0))
check_contrast_orthogonality(contrast_example)
}
\concept{stats}
