% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_out_path.R
\name{check_out_path}
\alias{check_out_path}
\title{check_out_path}
\usage{
check_out_path(out_path, overwrite = FALSE, skip = FALSE)
}
\arguments{
\item{out_path}{path to output file}

\item{overwrite}{booleon to overwrite the output path (default: FALSE)}

\item{skip}{boolean to skip if used in a for/while loop (default: FALSE)}
}
\value{

}
\description{
\preformatted{
checks to see if output file exists.

1. STOP WITH ERROR if output file exists, overwrite set to FALSE and skip to FALSE
2. WARN AND SKIP to next in the for/while loop if output file exists, overwrite set to FALSE, and skip set to TRUE
3. WARN if output file exists and overwrite set to TRUE
}
}
\examples{

}
\concept{r_helper}
