% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_paths.R
\name{check_out_paths}
\alias{check_out_paths}
\title{check_out_paths}
\usage{
check_out_paths(paths, overwrite = 0)
}
\arguments{
\item{paths}{paths to check}

\item{overwrite}{boolean}
}
\value{

}
\description{
check_out_paths
}
\concept{r_helper}
