% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_brains.R
\name{compare_brains}
\alias{compare_brains}
\title{compare_brains}
\usage{
compare_brains(a, b, mask, ci = 0.95, legend_title = NULL)
}
\arguments{
\item{a}{first image to compare}

\item{b}{second image to compare}

\item{mask}{mask image}

\item{ci}{confidence interval for bland-altman plots (default: 0.95)}
}
\value{

}
\description{
compare_brains
}
