% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_lut_fs_to_bids.R
\name{convet_lut_fs2bids}
\alias{convet_lut_fs2bids}
\title{convet_lut_fs2bids}
\usage{
convet_lut_fs2bids(lut, out_path = NULL, overwrite = FALSE)
}
\arguments{
\item{lut}{input path to look up table in freesurfer format}

\item{out_path}{output path to look up table in BIDS format (default: NULL)}

\item{overwrite}{flag to overwrite (default: 0)}
}
\value{
A data.frame with BIDS-style LUT columns; optionally writes it to \code{out_path}.
}
\description{
converts look up table (LUT) from freesurfer-format (No., Label Name, R, G, B, A) to BIDS-format (index, name, color [hex], alpha)
}
\seealso{
Other neuroimaging helpers:
\code{\link[=compare_brains]{compare_brains()}},
\code{\link[=extract_brain_slice]{extract_brain_slice()}},
\code{\link[=get_3dstats]{get_3dstats()}},
\code{\link[=plot_brain]{plot_brain()}},
\code{\link[=plot_brain_slice]{plot_brain_slice()}},
\code{\link[=split_dseg2mask]{split_dseg2mask()}}
}
\concept{neuroimaging}
\concept{neuroimaging helpers}
