% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2apa.R
\name{cor2apa}
\alias{cor2apa}
\title{cor2apa}
\usage{
cor2apa(m, digits = 3)
}
\arguments{
\item{m}{model from `cor.test()`}

\item{digits}{round statistics to specified digit}
}
\description{
converts `cor.test` model to APA format string
}
\examples{
cor.test()
}
