% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2apa.R
\name{cor2apa}
\alias{cor2apa}
\title{cor2apa}
\usage{
cor2apa(m, digits = 3)
}
\arguments{
\item{m}{model from \code{cor.test()}}

\item{digits}{round statistics to specified digit}
}
\description{
converts \code{cor.test()} model to APA format string
}
\examples{
cor2apa(cor.test(carData::Salaries$yrs.since.phd, carData::Salaries$salary))
}
\concept{stats}
