% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_kfold_split_file.R
\name{cv_kfold_split_file}
\alias{cv_kfold_split_file}
\title{cv_kfold_split_file}
\usage{
cv_kfold_split_file(x_path, y_path, k_fold, out_dir)
}
\arguments{
\item{out_dir}{}
}
\description{
cv_kfold_split_file
}
