% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_info.R
\name{error_msg}
\alias{error_msg}
\title{error_msg}
\usage{
error_msg(msg)
}
\arguments{
\item{msg}{error message to print}
}
\value{
Invisibly returns \code{NULL}; prints a formatted error message.
}
\description{
error_msg
}
\examples{
try(error_msg("example error"))
}
\seealso{
Other console messaging helpers: 
\code{\link{info_msg}()},
\code{\link{print_footer}()},
\code{\link{print_header}()},
\code{\link{warning_msg}()}
}
\concept{console messaging helpers}
