% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_cmd.R
\name{eval_cmd}
\alias{eval_cmd}
\title{eval_cmd}
\usage{
eval_cmd(cmd, out_paths, overwrite = FALSE, print = FALSE)
}
\arguments{
\item{cmd}{string expression to evaluate}

\item{out_paths}{vector or list of output paths}

\item{overwrite}{flag to overwrite output paths (default: FALSE)}

\item{print}{flag to print command only (default: FALSE)}
}
\value{

}
\description{
eval_cmd
}
\concept{r_helper}
