% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_dimension.R
\name{flatten_dimension}
\alias{flatten_dimension}
\title{flatten_dimension}
\usage{
flatten_dimension(data)
}
\arguments{
\item{data}{multi-dimensional data to be flattened/reduced by a single dimension}
}
\value{
data that is flattened/reduced by a single dimension
}
\description{
flatten_dimension
}
\examples{
# to be added
}
