% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_dimension_all.R
\name{flatten_dimension_all}
\alias{flatten_dimension_all}
\title{flatten_dimension_all}
\usage{
flatten_dimension_all(data)
}
\arguments{
\item{data}{multi-dimensional data to completely flattened/reduced to a single dimension}
}
\value{
data that is flattened to a single dimension
}
\description{
flatten_dimension_all
}
\examples{
# to be added
}
