% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_jn.R
\name{gen_data_jn}
\alias{gen_data_jn}
\title{Generate Johnson-Neyman Data}
\usage{
gen_data_jn(data, ci = 0.95)
}
\arguments{
\item{data}{A data.frame that is the output of \code{\link[=tidy_es]{tidy_es()}} on a \code{glht} object.}

\item{ci}{Confidence interval (0, 1). Default is 0.95.}
}
\value{
A data.frame of class \code{c("jn_df", "data.frame")} containing:
\itemize{
\item \code{x}: Values of the primary moderator (mapped to the x-axis).
\item \code{group}: Values of the secondary moderator, if any.
\item \code{facet}: Values of the tertiary moderator, if any.
\item \code{predicted}: Simple slope estimate.
\item \code{std.error}: Standard error of the slope.
\item \code{df}: Degrees of freedom.
\item \code{statistic}: t-statistic.
\item \code{p.value}: p-value.
\item \code{conf.low}: Lower bound of the slope confidence interval.
\item \code{conf.high}: Upper bound of the slope confidence interval.
\item \code{sig}: Logical indicating whether the slope is significant (p.value < 1 - ci).
}
}
\description{
Generate Johnson-Neyman Data
}
\examples{
fit1 <- lm(salary ~ yrs.since.phd * yrs.service, data = carData::Salaries)
c1 <- gen_contrast_ss(
  fit1,
  x = "yrs.since.phd",
  m = list(yrs.service = "real")
)
fit_cope1 <- multcomp::glht(fit1, linfct = c1)
df_cope_coef1 <- tidy_es(fit_cope1)
df_cope_jn1 <- gen_data_jn(df_cope_coef1)
plot(df_cope_jn1)
}
\seealso{
Other contrast or COPE helpers:
\code{\link[=check_contrast_orthogonality]{check_contrast_orthogonality()}},
\code{\link[=gen_contrast_blank]{gen_contrast_blank()}},
\code{\link[=gen_contrast_ss]{gen_contrast_ss()}},
\code{\link[=plot.jn_df]{plot.jn_df()}},
\code{\link[=tidy_es_cope_F]{tidy_es_cope_F()}},
\code{\link[=tidy_es_cope_t]{tidy_es_cope_t()}}
}
\concept{contrast or COPE helpers}
\concept{stats}
