% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_3dstats.R
\name{get_3dstats}
\alias{get_3dstats}
\title{get_3dstats}
\usage{
get_3dstats(in_data, in_dseg, in_lut = NULL, out_path, overwrite = FALSE)
}
\arguments{
\item{in_data}{input path to data nifti file}

\item{in_dseg}{input path to discrete segmentation nifti file}

\item{in_lut}{input path to look up table}

\item{out_path}{output path to save stats tsv file}

\item{overwrite}{flag to overwrite (default: 0)}
}
\description{
obtain summary metrics of data file from discrete segmentation. summary metrics include non-zero voxels, min, max, median, mean
}
\concept{neuroimaging}
