% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lm_multicollinearity.R
\name{get_lm_multicollinearity}
\alias{get_lm_multicollinearity}
\title{get_lm_multicollinearity}
\usage{
get_lm_multicollinearity(model)
}
\arguments{
\item{model}{fitted model from \code{lm()}}
}
\value{
A data.frame with multicollinearity diagnostics for each model term.
}
\description{
get_lm_multicollinearity
}
\examples{
get_lm_multicollinearity(lm(salary ~ yrs.since.phd + yrs.service, carData::Salaries))
}
\seealso{
Other model summary helpers: 
\code{\link{models2coefs}()},
\code{\link{models2omni}()},
\code{\link{r_sq_to_adj_r_sq}()},
\code{\link{renamed_tidy}()},
\code{\link{tidy_es_lmer}()},
\code{\link{tidy_es_lm}()},
\code{\link{tidy_es}()}
}
\concept{model summary helpers}
\concept{stats}
