% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lm_multicollinearity.R
\name{get_lm_multicollinearity}
\alias{get_lm_multicollinearity}
\title{get_lm_multicollinearity}
\usage{
get_lm_multicollinearity(model)
}
\arguments{
\item{model}{fitted model from `lm()`}
}
\description{
get_lm_multicollinearity
}
\concept{stats}
