% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_cope.R
\name{lm_cope_F}
\alias{lm_cope_F}
\title{lm_cope_F}
\usage{
lm_cope_F(model_fit, c, label)
}
\arguments{
\item{model_fit}{model fitted using \code{lm()}}

\item{c}{contrast matrix}

\item{label}{label for the contrast term}
}
\value{
data.frame with columns term, ss, df, ms, F, and p
}
\description{
perform a contrast over parameter estimate (COPE) \emph{F}-test
}
\examples{
model_fit <- lm(salary ~ rank + discipline, data = carData::Salaries)
c <- rbind(
  c(0, 1, 0, 0),
  c(0, 0, 1, 0)
)
lm_cope_F(model_fit, c, label = "rank")
}
\seealso{
Other lm_cope: 
\code{\link{lm_cope_t}()}
}
\concept{lm_cope}
\concept{stats}
