% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_cope.R
\name{lm_cope_t}
\alias{lm_cope_t}
\title{lm_cope_t}
\usage{
lm_cope_t(model_fit, c, ci = 0.95)
}
\arguments{
\item{model_fit}{model fitted using \code{lm()}}

\item{c}{contrast matrix}

\item{ci}{confidence interval (default: 0.95)}
}
\value{
data.frame with columns term, b, se, t, p, b_ci_ll, b_ci_ul, r_sq_adj, r_sq_adj_ci_ll, and r_sq_adj_ci_ul
}
\description{
perform a contrast over parameter estimate (COPE) \emph{t}-test
}
\examples{
model_fit <- lm(salary ~ rank + discipline, data = carData::Salaries)
c <- rbind(
  AssocProf_minus_AsstProf = c(0, 1, 0, 0),
  Prof_minus_AsstProf      = c(0, 0, 1, 0),
  Prof_minus_AssocProf     = c(0, -1, 1, 0)
)
lm_cope_t(model_fit, c)
}
\seealso{
Other lm_cope: 
\code{\link{lm_cope_F}()}
}
\concept{lm_cope}
\concept{stats}
