% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkdir_if_dne.R
\name{mkdir_if_dne}
\alias{mkdir_if_dne}
\title{mkdir_if_dne}
\usage{
mkdir_if_dne(dir, recursive = TRUE)
}
\arguments{
\item{dir}{}
}
\value{

}
\description{
recursively create directories if it does not exist
}
\concept{r_helper}
