% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_categorical_all.R
\name{model_categorical_all}
\alias{model_categorical_all}
\title{model_categorical_all}
\usage{
model_categorical_all(x, y)
}
\arguments{
\item{x}{features used in prediction}

\item{y}{outcome to predict}
}
\value{

}
\description{
model_categorical_all
}
\concept{model_uber}
