% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_categorical_elastic_net.R
\name{model_categorical_elastic_net}
\alias{model_categorical_elastic_net}
\title{model_categorical_elastic_net}
\usage{
model_categorical_elastic_net(x, y, alpha_list = seq(1e-04, 0.9999, 1e-04))
}
\arguments{
\item{x}{features/predictors used in prediction of outcome}

\item{y}{categorical outcome to predict}
}
\value{
elastic net results
}
\description{
model_categorical_elastic_net
}
\examples{
# to be added
}
\concept{model_categorical}
