% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_categorical_glm.R
\name{model_categorical_glm}
\alias{model_categorical_glm}
\title{model_categorical_glm}
\usage{
model_categorical_glm(x, y)
}
\arguments{
\item{x}{features/predictors used in prediction of outcome}

\item{y}{categorical outcome to predict}
}
\description{
model_categorical_glm
}
\concept{model_categorical}
