% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_categorical_lasso.R
\name{model_categorical_lasso}
\alias{model_categorical_lasso}
\title{model_categorical_lasso}
\usage{
model_categorical_lasso(x, y)
}
\arguments{
\item{x}{features/predictors used in prediction of outcome}

\item{y}{categorical outcome to predict}
}
\value{
lasso results
}
\description{
model_categorical_lasso
}
\examples{
# to be added
}
\concept{model_categorical}
