% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_categorical_ridge.R
\name{model_categorical_ridge}
\alias{model_categorical_ridge}
\title{model_categorical_ridge}
\usage{
model_categorical_ridge(x, y)
}
\arguments{
\item{x}{features/predictors used in prediction of outcome}

\item{y}{categorical outcome to predict}
}
\value{

}
\description{
model_categorical_ridge
}
\concept{model_categorical}
