% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_categorical_svm.R
\name{model_categorical_svm}
\alias{model_categorical_svm}
\title{model_categorical_svm}
\usage{
model_categorical_svm(x, y, kernel = "linear")
}
\arguments{
\item{x}{features/predictors used in prediction of outcome}

\item{y}{categorical outcome to predict}

\item{kernel}{}
}
\value{

}
\description{
model_categorical_svm
}
\concept{model_categorical}
