% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_continuous_elastic_net.R
\name{model_continous_elastic_net}
\alias{model_continous_elastic_net}
\title{model_continous_elastic_net}
\usage{
model_continous_elastic_net(data, y)
}
\arguments{
\item{data}{data to be analyzed}

\item{y}{name/string of outcome to be predicted that is within the data}
}
\value{
elastic net results
}
\description{
model_continous_elastic_net
}
\examples{
# to be added
}
