% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_continuous_ridge.R
\name{model_continuous_ridge}
\alias{model_continuous_ridge}
\title{model_continuous_ridge}
\usage{
model_continuous_ridge(data, y)
}
\arguments{
\item{data}{data to be analyzed}

\item{y}{name/string of outcome to be predicted that is within the data}
}
\value{
ridge results
}
\description{
model_continuous_ridge
}
\examples{
# to be added
}
