% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_continuous_svm_linear.R
\name{model_continuous_svm_linear}
\alias{model_continuous_svm_linear}
\title{model_continuous_svm_linear}
\usage{
model_continuous_svm_linear(data, y)
}
\arguments{
\item{data}{clean data.frame of predictors and outcome to be analyzed}

\item{y}{continuous outcome name within the data.frame}
}
\value{
linear svm model
}
\description{
model continuous linear support vector machine
}
\examples{
# to be added
}
