% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models2omni.R
\name{models2omni}
\alias{models2omni}
\title{models2omni}
\usage{
models2omni(models, attribute = "extra_info")
}
\arguments{
\item{models}{a single model or a list of models}

\item{attribute}{name of attribute list to turn in columns (default: 'extra_info')}
}
\value{
data.frame of extra_info of model attributes and glance
}
\description{
model(s) into a single omnibus table. Uses \code{broom::glance()} and adds extra columns of extra model attributes
}
\examples{
m1 <- lm(salary ~ yrs.since.phd, carData::Salaries)
attr(m1, 'extra_info') <- list(m = 1, another_key = 'another_value')
models2omni(m1)
}
\seealso{
Other model summary helpers:
\code{\link[=get_lm_multicollinearity]{get_lm_multicollinearity()}},
\code{\link[=models2coefs]{models2coefs()}},
\code{\link[=r_sq_to_adj_r_sq]{r_sq_to_adj_r_sq()}},
\code{\link[=renamed_tidy]{renamed_tidy()}},
\code{\link[=tidy_es]{tidy_es()}},
\code{\link[=tidy_es_lm]{tidy_es_lm()}},
\code{\link[=tidy_es_lmer]{tidy_es_lmer()}}
}
\concept{model summary helpers}
\concept{stats}
