% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_table.R
\name{nice_table}
\alias{nice_table}
\title{nice_table}
\usage{
nice_table(table)
}
\arguments{
\item{table}{a data.frame/table to pass through kable}
}
\value{
a kable-styled table
}
\description{
nice_table
}
