% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_table_html_short.R
\name{nice_table_html_short}
\alias{nice_table_html_short}
\title{nice_table_html_short}
\usage{
nice_table_html_short(table, height = "200px")
}
\arguments{
\item{height}{}
}
\description{
nice_table_html_short
}
