% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_jn.R
\name{plot.jn_df}
\alias{plot.jn_df}
\title{Plot Johnson-Neyman Data}
\usage{
\method{plot}{jn_df}(x, scales = "fixed", ...)
}
\arguments{
\item{x}{An object of class \code{jn_df} (the output from \code{\link[=gen_data_jn]{gen_data_jn()}}).}

\item{scales}{Character vector specifying whether facet scales should be \code{"fixed"}, \code{"free_y"}, \code{"free_x"}, or \code{"free"}. Default is \code{"fixed"}.}

\item{...}{Additional arguments (not used).}
}
\value{
A \code{ggplot} object.
}
\description{
Plot Johnson-Neyman Data
}
\seealso{
Other contrast or COPE helpers:
\code{\link[=check_contrast_orthogonality]{check_contrast_orthogonality()}},
\code{\link[=gen_contrast_blank]{gen_contrast_blank()}},
\code{\link[=gen_contrast_ss]{gen_contrast_ss()}},
\code{\link[=gen_data_jn]{gen_data_jn()}},
\code{\link[=tidy_es_cope_F]{tidy_es_cope_F()}},
\code{\link[=tidy_es_cope_t]{tidy_es_cope_t()}}
}
\concept{contrast or COPE helpers}
\concept{stats}
