% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age.R
\name{plot_age}
\alias{plot_age}
\title{plot_age}
\usage{
plot_age(data, y_var, x_var = "age", y_lab = NULL, x_lab = "Age")
}
\arguments{
\item{data}{input data.frame}

\item{y_lab}{y-axis label}

\item{x_lab}{x-axis label}

\item{var_y}{y-axis variable to plot}

\item{var_x}{x-axis variable to plot}
}
\value{
ggplot figure of y-axis variable and age
}
\description{
plot_age
}
\concept{visualization}
