% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age.R
\name{plot_age}
\alias{plot_age}
\title{plot_age}
\usage{
plot_age(data, y_var, x_var = "age", y_lab = NULL, x_lab = "Age")
}
\arguments{
\item{data}{}

\item{y_lab}{}

\item{var_y}{}

\item{var_x}{}
}
\description{
plot_age
}
