% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cor.R
\name{plot_cor}
\alias{plot_cor}
\title{plot_cor}
\usage{
plot_cor(x, y, data)
}
\arguments{
\item{x}{x variable}

\item{y}{y variable}

\item{data}{data that contains the x and y variables}
}
\value{
ggplot2 figure with correlation statistics
}
\description{
plot_cor
}
\examples{
plot_cor('yrs.service', 'yrs.since.phd', carData::Salaries)
}
\concept{visualization}
