% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cor_heatmap.R
\name{plot_cor_heatmap}
\alias{plot_cor_heatmap}
\title{plot_cor_heatmap}
\usage{
plot_cor_heatmap(data)
}
\arguments{
\item{data}{data to create a heat map from}
}
\value{
ggplot figure
}
\description{
plot_cor_heatmap
}
\examples{
plot_cor_heatmap(mtcars)
}
\concept{visualization}
