% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_steps.R
\name{plot_steps}
\alias{plot_steps}
\title{plot_steps}
\usage{
plot_steps(
  in_path,
  out_path = NULL,
  subjid = "sub",
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{in_path}{path to input csv file of subjects by rows (required)}

\item{out_path}{path to save output image (optional)}

\item{subjid}{columns that identify the rows. if more than one, will combine into key1-value1_key2-value2_...keyN-valueN}

\item{title}{title of plot}

\item{subtitle}{subtitle of plot}
}
\value{
ggplot2 heatmap of steps
}
\description{
plot_steps
}
\concept{visualization}
