% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmid2funcs.R
\name{pmid2abstract}
\alias{pmid2abstract}
\title{pmid2abstract}
\usage{
pmid2abstract(pmid)
}
\arguments{
\item{pmid}{}
}
\description{
pmid2abstract
}
\examples{
pmid2abstract(33317393)
}
\concept{references_pubmed}
