% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmid2funcs.R
\name{pmid2bibtex}
\alias{pmid2bibtex}
\title{pmid2bibtex}
\usage{
pmid2bibtex(pmid, key_as_first_author_year = TRUE)
}
\arguments{
\item{key_as_first_author_year}{}
}
\description{
pmid2bibtex
}
\concept{references_pubmed}
