% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_sq_to_adj_r_sq.R
\name{r_sq_to_adj_r_sq}
\alias{r_sq_to_adj_r_sq}
\title{r_sq_to_adj_r_sq}
\usage{
r_sq_to_adj_r_sq(r_sq, n, n_p)
}
\arguments{
\item{r_sq}{r-squared (r^2) value}

\item{n}{number of subjects/rows}

\item{n_p}{number of parameters}
}
\value{
adjusted r-squared (r^2) value
}
\description{
r_sq_to_adj_r_sq
}
\examples{
# to be added
}
