% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_sq_to_adj_r_sq.R
\name{r_sq_to_adj_r_sq}
\alias{r_sq_to_adj_r_sq}
\title{r_sq_to_adj_r_sq}
\usage{
r_sq_to_adj_r_sq(r_sq, n, n_p)
}
\arguments{
\item{r_sq}{r-squared ($R^2$) value}

\item{n}{number of subjects/rows}

\item{n_p}{number of parameters}
}
\value{

}
\description{
Convert r-squared ($R^2$) to adjusted r-squared ($R^2_{Adj.}$).
}
\examples{
r_sq_to_adj_r_sq(.5, 100, 5)
}
\concept{stats}
