% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_colnames.R
\name{rename_colnames}
\alias{rename_colnames}
\title{rename_colnames}
\usage{
rename_colnames(data, new_colnames)
}
\arguments{
\item{data}{data.frame to be renamed}

\item{new_colnames}{a vector of new column names}
}
\value{
data.frame with the new column names
}
\description{
rename column names
}
\examples{
# example to be added
}
