% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_to_write.R
\name{safe_to_write}
\alias{safe_to_write}
\title{safe_to_write}
\usage{
safe_to_write(f, overwrite = FALSE)
}
\arguments{
\item{f}{path to file}

\item{overwrite}{logical, whether to overwrite if file already exists (default: FALSE)}
}
\value{
logical, whether it is safe to write to the file
}
\description{
safe_to_write
}
