% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dseg2mask.R
\name{split_dseg2mask}
\alias{split_dseg2mask}
\title{split_dseg2mask}
\usage{
split_dseg2mask(in_dseg, in_lut, out_prefix, overwrite = FALSE)
}
\arguments{
\item{in_dseg}{path to discrete segmentation file}

\item{in_lut}{path to look up table .tsv file (bids format)}

\item{out_prefix}{output prefix of the mask files}

\item{overwrite}{flag to overwrite output files (default: FALSE)}
}
\value{
Invisibly returns \code{NULL}; writes one mask NIfTI per LUT entry.
}
\description{
split_dseg2mask
}
\seealso{
Other neuroimaging helpers: 
\code{\link{compare_brains}()},
\code{\link{convet_lut_fs2bids}()},
\code{\link{extract_brain_slice}()},
\code{\link{get_3dstats}()},
\code{\link{plot_brain_slice}()},
\code{\link{plot_brain}()}
}
\concept{neuroimaging}
\concept{neuroimaging helpers}
