% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_es.R
\name{tidy_es}
\alias{tidy_es}
\title{tidy_es}
\usage{
tidy_es(
  model,
  ci = 0.95,
  mcc = c("single-step", "none", "uncorrected", "fdr", "FDR", "tukey", "tukey's HSD",
    "Tukey")
)
}
\arguments{
\item{model}{model fit from \code{lm()}, \code{lmerTest::lmer()}, \code{lme4::lmer()}, or \code{multcomp::glht()}}

\item{ci}{confidence interval (0, 1)}

\item{mcc}{multiple comparison correction method. Options include "single-step" (default, which is what glht provides for Tukey's HSD), "none" (uncorrected), "fdr", "FDR", "tukey", "tukey's HSD", and "Tukey".}
}
\value{
data.frame
}
\description{
a modification of the renamed tidy command to adds the effect
size (confidence intervals for the regression coefficient and adjusted R^2)
}
\examples{
# example: lm
tidy_es(lm(salary ~ yrs.since.phd, carData::Salaries))

# example: lmer
tidy_es(lmerTest::lmer(Reaction ~ 1 + Days + (1 + Days | Subject), lme4::sleepstudy))

# example: glht
model_fit <- lm(salary ~ rank + discipline, carData::Salaries)
c <- rbind(
  AssocProf_minus_AsstProf = c(0, 1, 0, 0),
  Prof_minus_AsstProf      = c(0, 0, 1, 0),
  Prof_minus_AssocProf     = c(0, -1, 1, 0)
)
model_fit_cope_t <- multcomp::glht(model_fit, c)
tidy_es(model_fit_cope_t)
}
\seealso{
Other model summary helpers:
\code{\link[=get_lm_multicollinearity]{get_lm_multicollinearity()}},
\code{\link[=models2coefs]{models2coefs()}},
\code{\link[=models2omni]{models2omni()}},
\code{\link[=r_sq_to_adj_r_sq]{r_sq_to_adj_r_sq()}},
\code{\link[=renamed_tidy]{renamed_tidy()}},
\code{\link[=tidy_es_cope_F]{tidy_es_cope_F()}},
\code{\link[=tidy_es_cope_t]{tidy_es_cope_t()}},
\code{\link[=tidy_es_lm]{tidy_es_lm()}},
\code{\link[=tidy_es_lmer]{tidy_es_lmer()}}
}
\concept{model summary helpers}
\concept{stats}
