% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_es.R
\name{tidy_es}
\alias{tidy_es}
\title{tidy_es}
\usage{
tidy_es(model, ci = 0.95)
}
\arguments{
\item{model}{model fit from lm()}

\item{ci}{confidence interval (0, 1)}
}
\value{
data.frame
}
\description{
a modification of the tidy command
}
\examples{
tidy_eep(salary ~ yrs.since.phd, carData::Salaries)
}
