% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_es.R
\name{tidy_es_cope_F}
\alias{tidy_es_cope_F}
\title{tidy_es_cope_F}
\usage{
tidy_es_cope_F(model, label = NULL, ci = 0.95)
}
\arguments{
\item{model}{contrast model from \code{multcomp::glht()}}

\item{label}{label for the contrast term (default: joined contrast names)}

\item{ci}{confidence interval (0, 1)}
}
\value{
data.frame
}
\description{
a modification of the tidy command that adds the effect
size (confidence intervals for the adjusted R^2) for linear hypothesis
testing contrasts (joint F-test) using glht
}
\examples{
model_fit <- lm(salary ~ rank + discipline, data = carData::Salaries)
c <- gen_contrast_ss(model_fit, x = "rank")[-1, ]
model_fit_cope_F <- multcomp::glht(model_fit, c)
tidy_es_cope_F(model_fit_cope_F, label = "rank")
}
\seealso{
Other model summary helpers:
\code{\link[=get_lm_multicollinearity]{get_lm_multicollinearity()}},
\code{\link[=models2coefs]{models2coefs()}},
\code{\link[=models2omni]{models2omni()}},
\code{\link[=r_sq_to_adj_r_sq]{r_sq_to_adj_r_sq()}},
\code{\link[=renamed_tidy]{renamed_tidy()}},
\code{\link[=tidy_es]{tidy_es()}},
\code{\link[=tidy_es_cope_t]{tidy_es_cope_t()}},
\code{\link[=tidy_es_lm]{tidy_es_lm()}},
\code{\link[=tidy_es_lmer]{tidy_es_lmer()}}

Other contrast or COPE helpers:
\code{\link[=check_contrast_orthogonality]{check_contrast_orthogonality()}},
\code{\link[=gen_contrast_blank]{gen_contrast_blank()}},
\code{\link[=gen_contrast_ss]{gen_contrast_ss()}},
\code{\link[=tidy_es_cope_t]{tidy_es_cope_t()}}
}
\concept{contrast or COPE helpers}
\concept{model summary helpers}
\concept{stats}
