% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_es.R
\name{tidy_es_cope_t}
\alias{tidy_es_cope_t}
\title{tidy_es_cope_t}
\usage{
tidy_es_cope_t(model, ci = 0.95)
}
\arguments{
\item{model}{contrast model from \code{multcomp::glht()}}

\item{ci}{confidence interval (0, 1)}
}
\value{
data.frame
}
\description{
a modification of the tidy command that adds the effect
size (confidence intervals for the contrast coefficient and adjusted R^2)
for linear hypothesis testing contrasts (t-test) using glht
}
\examples{
model_fit <- lm(salary ~ rank + discipline, data = carData::Salaries)
c <- rbind(
  AssocProf_minus_AsstProf = c(0, 1, 0, 0),
  Prof_minus_AsstProf      = c(0, 0, 1, 0),
  Prof_minus_AssocProf     = c(0, -1, 1, 0)
)
model_fit_cope_t <- multcomp::glht(model_fit, c)
tidy_es_cope_t(model_fit_cope_t)
}
\seealso{
Other model summary helpers:
\code{\link[=get_lm_multicollinearity]{get_lm_multicollinearity()}},
\code{\link[=models2coefs]{models2coefs()}},
\code{\link[=models2omni]{models2omni()}},
\code{\link[=r_sq_to_adj_r_sq]{r_sq_to_adj_r_sq()}},
\code{\link[=renamed_tidy]{renamed_tidy()}},
\code{\link[=tidy_es]{tidy_es()}},
\code{\link[=tidy_es_cope_F]{tidy_es_cope_F()}},
\code{\link[=tidy_es_lm]{tidy_es_lm()}},
\code{\link[=tidy_es_lmer]{tidy_es_lmer()}}
}
\concept{model summary helpers}
\concept{stats}
