% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_es.R
\name{tidy_es_lm}
\alias{tidy_es_lm}
\title{tidy_es_lm}
\usage{
tidy_es_lm(model, ci = 0.95)
}
\arguments{
\item{model}{model fit from \code{lm()}}

\item{ci}{confidence interval (0, 1)}
}
\value{
data.frame
}
\description{
a modification of the renamed tidy command that adds the effect
size (confidence intervals for the regression coefficient and adjusted R^2)
}
\examples{
tidy_es_lm(lm(salary ~ yrs.since.phd, carData::Salaries))
}
\concept{stats}
