% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_es.R
\name{tidy_es_lmer}
\alias{tidy_es_lmer}
\title{tidy_es_lmer}
\usage{
tidy_es_lmer(model, ci = 0.95)
}
\arguments{
\item{model}{model output from \code{lmerTest::lmer()} or \code{lme4::lmer()}}

\item{ci}{confidence interval (default: 0.95)}
}
\value{
data.frame
}
\description{
a modification of the renamed tidy command that adds the effect
size (confidence intervals for the regression coefficient and adjusted R^2)
}
\examples{
tidy_es_lmer(lmerTest::lmer(Reaction ~ 1 + Days + (1 + Days | Subject), lme4::sleepstudy))
}
\concept{stats}
